+++ app/src/main/java/uz/orifjon/educationsysteminandroid/models/Course.kt
@@ -1,3 +1,3 @@
 package uz.orifjon.educationsysteminandroid.models
 
-data class Course(var id: Int = 0, var name: String, var description: String)
+data class Course(var id: Long = 0, var name: String, var description: String)
