+++ app/src/main/java/uz/orifjon/educationsysteminandroid/models/Student.kt
@@ -1,3 +1,3 @@
 package uz.orifjon.educationsysteminandroid.models
 
-data class Student(val id: Int = 0,var firstname:String,var lastname:String,var patron:String,var registerDate:String,var groupId:Int)
+data class Student(val id: Long = 0,var firstname:String,var lastname:String,var patron:String,var registerDate:String,var groupId:Int)
